requirements.txt
fastmcp
playwright
python-dotenv
PyYAML
openai
streamlit
pytest
pytest-mock 